This revision is a large update to add support for keyword searches as suggested by Rutt on the Slim Devices forums. The code still needs to be tidied up - this is just out as a beta for some validation before I release it wider.
This has involved significant change to the code in most areas, and such the existing lazy search functionality may have been destabilised and so also needs re-testing.

There is now an additional top-level lazy search menu item for keyword searches, that appears after the existing entry for song searches.

Using keyword searches should be pretty self-explanatory. Multiple words are entered separated by the '0' key, and the results are filtered to those that match all keywords. The results are presented in a menu of ARTISTS, ALBUMS, and SONGS. As an example, the keyword search "4732804487" will look for the words "GREAT" and "HITS" which, in my case, just finds lots of greatest hits albums.

The minimum length of each keyword is four characters by default, but that can be changed through "Server Settings-Plugins" as for the other defaults. Also, you can press SEARCH to force a search even with short keywords (if there are at least two characters).

Other changes:

* Fixed a bug with the use of INPUT.Choice - in the past the user may have not entered the list at item 1 if other search lists had been looked at previously.

Still to do:

* Remove some low-level debugging log messages.

* Translations for the new strings.

* Prettify the code.