Updates to the keyword search mechanism after feedback from Rutt. Search now proceeds as follows:
-------
ARTISTS
-------
All artists with a name that matches any keyword AND
	(
		have an album with a title that matches any keyword
	OR
		have a track with a title that matches any keyword
	)

------
ALBUMS
------
All albums with a name that matches any keyword AND
	(
		by an artist with a name that matches any keyword
	OR
		contain a track with a title that matches any keyword
	)

------
TRACKS
------
All tracks with a name that matches any keyword AND
	(
		by an artist with a name that matches any keyword
	OR
		appear on an album with a name that matches any keyword
	)

Note that this generally means that you get no search results until you've put at least two keywords in (separated by '0'), since otherwise it's looking for artists having albums with the same word in the title etc.

This is still only a beta release, and still needs the original searching method given a good degree of testing as that has had to change a lot to accommodate the keyword search mechanism.