Third go at the keyword search queries; this time the search keywords are applied across all the artist/album/song and the results presented in the standard browse hierarchy.
As before, keywords are AND'd together, and the '0' key on the remote control is used to separate each keyword. For example, entering '9255035693' would find items matching both WALL and FLOYD.

This still has a lot of extra debugging output in it that needs to be taken out, and the source needs to be re-prettified.

Note that the lazified database format has changed and this means the database needs to be re-lazified. The plugin will do this automatically the first time it is started (so no manual rescan should be necessary), but be aware that keyword searches won't find all the correct results until that process has completed.

The plugin now mentions both Squeezebox and Transporter as players.