Fixed MusicIP mixes for keyword searches
ks10 reported on the forums that keyword searches wouldn't allow MusiCIP
mixes to be created (it tried, but always failed). True enough it seems
this has never worked.

Problem was that the mix type wasn't being set in our custom mode and so
we didn't know which type to request from the MusicIP plugin when the
user requested that function.