Fix playing of tracks in albums
Now, when "play other tracks in album" preference is set for the player
or server and the user searches and plays a single track, the whole
album is added to the playlist and playing begins from the track the
original search was on.

This behaviour is then consistent with the standard search when a track
is chosen from an album.